Begin making materials generic
