+++ rend3/src/lib.rs
@@ -76,6 +76,7 @@ pub mod util {
     pub mod bind_merge;
     pub mod buffer;
     pub mod frustum;
+    pub mod math;
     pub mod mipmap;
     pub mod output;
     pub mod registry;
