+++ rend3/src/resources/material.rs
@@ -1,292 +1,55 @@
-use crate::{
-    mode::ModeData,
-    resources::TextureManager,
-    types::{Material, MaterialChange, MaterialFlags, MaterialHandle, SampleType, TextureHandle},
-    util::{bind_merge::BindGroupBuilder, buffer::WrappedPotBuffer, registry::ResourceRegistry},
-    RendererMode,
-};
+use crate::{RendererMode, mode::ModeData, resources::TextureManager, types::{Material, MaterialChange, MaterialFlags, MaterialHandle, SampleType, TextureHandle}, util::{bind_merge::BindGroupBuilder, buffer::WrappedPotBuffer, math::round_up_pot, registry::{ArchitypeResourceStorage, ArchitypicalRegistry, ResourceRegistry}, typedefs::FastHashMap}};
 use glam::{Vec3, Vec4};
-use rend3_types::RawMaterialHandle;
+use list_any::VecAny;
+use rend3_types::{MaterialTrait, RawMaterialHandle};
 use std::{
+    any::TypeId,
     mem,
     num::{NonZeroU32, NonZeroU64},
 };
 use wgpu::{
     util::{BufferInitDescriptor, DeviceExt},
-    BindGroup, BindGroupLayout, BindGroupLayoutDescriptor, BindGroupLayoutEntry, BindingType, Buffer,
+    BindGroup, BindGroupLayout, BindGroupLayoutDescriptor, BindGroupLayoutEntry, BindingResource, BindingType, Buffer,
     BufferBindingType, BufferUsages, Device, Queue, ShaderStages, TextureSampleType, TextureViewDimension,
 };
 
-#[repr(C, align(16))]
-#[derive(Debug, Copy, Clone)]
-struct CPUShaderMaterial {
-    uv_transform0_row0: Vec4,
-    uv_transform0_row1: Vec4,
-    uv_transform0_row2: Vec4,
-    uv_transform1_row0: Vec4,
-    uv_transform1_row1: Vec4,
-    uv_transform1_row2: Vec4,
-    albedo: Vec4,
-    emissive: Vec3,
-    roughness: f32,
-    metallic: f32,
-    reflectance: f32,
-    clear_coat: f32,
-    clear_coat_roughness: f32,
-    anisotropy: f32,
-    ambient_occlusion: f32,
-    alpha_cutout: f32,
-
-    texture_enable: u32,
-    material_flags: MaterialFlags,
-}
-
-unsafe impl bytemuck::Zeroable for CPUShaderMaterial {}
-unsafe impl bytemuck::Pod for CPUShaderMaterial {}
-
-impl CPUShaderMaterial {
-    fn from_material(material: &Material) -> Self {
-        Self {
-            uv_transform0_row0: material.uv_transform0.x_axis.extend(0.0),
-            uv_transform0_row1: material.uv_transform0.y_axis.extend(0.0),
-            uv_transform0_row2: material.uv_transform0.z_axis.extend(0.0),
-            uv_transform1_row0: material.uv_transform1.x_axis.extend(0.0),
-            uv_transform1_row1: material.uv_transform1.y_axis.extend(0.0),
-            uv_transform1_row2: material.uv_transform1.z_axis.extend(0.0),
-            albedo: material.albedo.to_value(),
-            roughness: material.roughness_factor.unwrap_or(0.0),
-            metallic: material.metallic_factor.unwrap_or(0.0),
-            reflectance: material.reflectance.to_value(0.5),
-            clear_coat: material.clearcoat_factor.unwrap_or(0.0),
-            clear_coat_roughness: material.clearcoat_roughness_factor.unwrap_or(0.0),
-            emissive: material.emissive.to_value(Vec3::ZERO),
-            anisotropy: material.anisotropy.to_value(0.0),
-            ambient_occlusion: material.ao_factor.unwrap_or(1.0),
-            alpha_cutout: match material.transparency {
-                rend3_types::Transparency::Cutout { cutout } => cutout,
-                _ => 0.0,
-            },
-            texture_enable: material.albedo.is_texture() as u32
-                | (material.normal.to_texture(|_| ()).is_some() as u32) << 1
-                | (material.aomr_textures.to_roughness_texture(|_| ()).is_some() as u32) << 2
-                | (material.aomr_textures.to_metallic_texture(|_| ()).is_some() as u32) << 3
-                | (material.reflectance.is_texture() as u32) << 4
-                | (material.clearcoat_textures.to_clearcoat_texture(|_| ()).is_some() as u32) << 5
-                | (material
-                    .clearcoat_textures
-                    .to_clearcoat_roughness_texture(|_| ())
-                    .is_some() as u32)
-                    << 6
-                | (material.emissive.is_texture() as u32) << 7
-                | (material.anisotropy.is_texture() as u32) << 8
-                | (material.aomr_textures.to_ao_texture(|_| ()).is_some() as u32) << 9,
-            material_flags: {
-                let mut flags = material.albedo.to_flags();
-                flags |= material.normal.to_flags();
-                flags |= material.aomr_textures.to_flags();
-                flags |= material.clearcoat_textures.to_flags();
-                flags.set(MaterialFlags::UNLIT, material.unlit);
-                flags.set(
-                    MaterialFlags::NEAREST,
-                    match material.sample_type {
-                        SampleType::Nearest => true,
-                        SampleType::Linear => false,
-                    },
-                );
-                flags
-            },
-        }
-    }
-}
-
-#[repr(C, align(16))]
-#[derive(Debug, Copy, Clone)]
-struct GPUShaderMaterial {
-    albedo: Vec4,
-    emissive: Vec3,
-    roughness: f32,
-    metallic: f32,
-    reflectance: f32,
-    clear_coat: f32,
-    clear_coat_roughness: f32,
-    anisotropy: f32,
-    ambient_occlusion: f32,
-    alpha_cutout: f32,
-
-    uv_transform0_row0: Vec4,
-    uv_transform0_row1: Vec4,
-    uv_transform0_row2: Vec4,
-
-    uv_transform1_row0: Vec4,
-    uv_transform1_row1: Vec4,
-    uv_transform1_row2: Vec4,
-
-    albedo_tex: Option<NonZeroU32>,
-    normal_tex: Option<NonZeroU32>,
-    roughness_tex: Option<NonZeroU32>,
-    metallic_tex: Option<NonZeroU32>,
-    reflectance_tex: Option<NonZeroU32>,
-    clear_coat_tex: Option<NonZeroU32>,
-    clear_coat_roughness_tex: Option<NonZeroU32>,
-    emissive_tex: Option<NonZeroU32>,
-    anisotropy_tex: Option<NonZeroU32>,
-    ambient_occlusion_tex: Option<NonZeroU32>,
-    material_flags: MaterialFlags,
-}
-
-unsafe impl bytemuck::Zeroable for GPUShaderMaterial {}
-unsafe impl bytemuck::Pod for GPUShaderMaterial {}
-
-impl GPUShaderMaterial {
-    pub fn from_material(material: &Material, translate_texture: &impl Fn(&TextureHandle) -> NonZeroU32) -> Self {
-        Self {
-            albedo: material.albedo.to_value(),
-            emissive: material.emissive.to_value(Vec3::ZERO),
-            roughness: material.roughness_factor.unwrap_or(0.0),
-            metallic: material.metallic_factor.unwrap_or(0.0),
-            reflectance: material.reflectance.to_value(0.5),
-            clear_coat: material.clearcoat_factor.unwrap_or(0.0),
-            clear_coat_roughness: material.clearcoat_roughness_factor.unwrap_or(0.0),
-            anisotropy: material.anisotropy.to_value(0.0),
-            ambient_occlusion: material.ao_factor.unwrap_or(1.0),
-            alpha_cutout: match material.transparency {
-                rend3_types::Transparency::Cutout { cutout } => cutout,
-                _ => 0.0,
-            },
-
-            uv_transform0_row0: material.uv_transform0.x_axis.extend(0.0),
-            uv_transform0_row1: material.uv_transform0.y_axis.extend(0.0),
-            uv_transform0_row2: material.uv_transform0.z_axis.extend(0.0),
-
-            uv_transform1_row0: material.uv_transform1.x_axis.extend(0.0),
-            uv_transform1_row1: material.uv_transform1.y_axis.extend(0.0),
-            uv_transform1_row2: material.uv_transform1.z_axis.extend(0.0),
-
-            albedo_tex: material.albedo.to_texture(translate_texture),
-            normal_tex: material.normal.to_texture(translate_texture),
-            roughness_tex: material.aomr_textures.to_roughness_texture(translate_texture),
-            metallic_tex: material.aomr_textures.to_metallic_texture(translate_texture),
-            reflectance_tex: material.reflectance.to_texture(translate_texture),
-            clear_coat_tex: material.clearcoat_textures.to_clearcoat_texture(translate_texture),
-            clear_coat_roughness_tex: material
-                .clearcoat_textures
-                .to_clearcoat_roughness_texture(translate_texture),
-            emissive_tex: material.emissive.to_texture(translate_texture),
-            anisotropy_tex: material.anisotropy.to_texture(translate_texture),
-            ambient_occlusion_tex: material.aomr_textures.to_ao_texture(translate_texture),
-            material_flags: {
-                let mut flags = material.albedo.to_flags();
-                flags |= material.normal.to_flags();
-                flags |= material.aomr_textures.to_flags();
-                flags |= material.clearcoat_textures.to_flags();
-                flags.set(MaterialFlags::UNLIT, material.unlit);
-                flags.set(
-                    MaterialFlags::NEAREST,
-                    match material.sample_type {
-                        SampleType::Nearest => true,
-                        SampleType::Linear => false,
-                    },
-                );
-                flags
-            },
-        }
-    }
-}
-
-struct InternalMaterial {
-    mat: Material,
+struct InternalMaterial<M: MaterialTrait> {
+    mat: M,
     bind_group: ModeData<BindGroup, ()>,
     material_buffer: ModeData<Buffer, ()>,
 }
 
+#[derive(Copy, Clone)]
+struct PerTypeInfo {
+    data_count: u32,
+    texture_count: u32,
+}
+
 /// Manages materials and their associated BindGroups in CPU modes.
 pub struct MaterialManager {
-    bgl: ModeData<BindGroupLayout, BindGroupLayout>,
-    bg: ModeData<(), BindGroup>,
+    bgl: FastHashMap<TypeId, ModeData<BindGroupLayout, BindGroupLayout>>,
+    bg: FastHashMap<TypeId, ModeData<(), BindGroup>>,
+
+    type_info: FastHashMap<TypeId, PerTypeInfo>,
+
     buffer: ModeData<(), WrappedPotBuffer>,
 
-    registry: ResourceRegistry<InternalMaterial, Material>,
+    registry: ArchitypicalRegistry<Material>,
 }
 
 impl MaterialManager {
     pub fn new(device: &Device, mode: RendererMode) -> Self {
-        let bgl = mode.into_data(
-            || {
-                let texture_binding = |idx| BindGroupLayoutEntry {
-                    binding: idx,
-                    visibility: ShaderStages::FRAGMENT,
-                    ty: BindingType::Texture {
-                        sample_type: TextureSampleType::Float { filterable: true },
-                        view_dimension: TextureViewDimension::D2,
-                        multisampled: false,
-                    },
-                    count: None,
-                };
-
-                device.create_bind_group_layout(&BindGroupLayoutDescriptor {
-                    label: Some("cpu material bgl"),
-                    entries: &[
-                        texture_binding(0),
-                        texture_binding(1),
-                        texture_binding(2),
-                        texture_binding(3),
-                        texture_binding(4),
-                        texture_binding(5),
-                        texture_binding(6),
-                        texture_binding(7),
-                        texture_binding(8),
-                        texture_binding(9),
-                        BindGroupLayoutEntry {
-                            binding: 10,
-                            visibility: ShaderStages::VERTEX_FRAGMENT,
-                            ty: BindingType::Buffer {
-                                ty: BufferBindingType::Uniform,
-                                has_dynamic_offset: false,
-                                min_binding_size: NonZeroU64::new(mem::size_of::<CPUShaderMaterial>() as _),
-                            },
-                            count: None,
-                        },
-                    ],
-                })
-            },
-            || {
-                device.create_bind_group_layout(&BindGroupLayoutDescriptor {
-                    label: Some("gpu material bgl"),
-                    entries: &[BindGroupLayoutEntry {
-                        binding: 0,
-                        visibility: ShaderStages::VERTEX_FRAGMENT,
-                        ty: BindingType::Buffer {
-                            ty: BufferBindingType::Storage { read_only: true },
-                            has_dynamic_offset: false,
-                            min_binding_size: NonZeroU64::new(mem::size_of::<GPUShaderMaterial>() as _),
-                        },
-                        count: None,
-                    }],
-                })
-            },
-        );
-
         let buffer = mode.into_data(
             || (),
-            || {
-                WrappedPotBuffer::new(
-                    device,
-                    0,
-                    mem::size_of::<GPUShaderMaterial>() as _,
-                    BufferUsages::STORAGE,
-                    Some("material buffer"),
-                )
-            },
+            || WrappedPotBuffer::new(device, 0, 16 as _, BufferUsages::STORAGE, Some("material buffer")),
         );
 
-        let bg = mode.into_data(|| (), || create_gpu_buffer_bg(device, bgl.as_gpu(), buffer.as_gpu()));
-
-        let registry = ResourceRegistry::new();
+        let registry = ArchitypicalRegistry::new();
 
         Self {
-            bgl,
-            bg,
-
+            bgl: FastHashMap::default(),
+            bg: FastHashMap::default(),
+            type_info: FastHashMap::default(),
             buffer,
             registry,
         }
@@ -296,69 +59,115 @@ impl MaterialManager {
         self.registry.allocate()
     }
 
-    pub fn fill(
+    pub fn fill<M: MaterialTrait>(
         &mut self,
         device: &Device,
         mode: RendererMode,
         texture_manager_2d: &mut TextureManager,
         handle: &MaterialHandle,
-        material: Material,
+        material: M,
     ) {
         let null_tex = texture_manager_2d.get_null_view();
 
+        let texture_count = material.texture_count();
+        let data_count = material.data_count();
+
         let material_buffer = mode.into_data(
             || {
-                let data = CPUShaderMaterial::from_material(&material);
+                // TODO: stack allocation
+                let mut data = vec![0u8; data_count as usize];
+                material.to_data(&mut data);
 
                 device.create_buffer_init(&BufferInitDescriptor {
                     label: None,
-                    contents: bytemuck::bytes_of(&data),
+                    contents: &data,
                     usage: BufferUsages::COPY_DST | BufferUsages::UNIFORM,
                 })
             },
             || (),
         );
 
-        let lookup_fn = |handle: &TextureHandle| texture_manager_2d.get_view(handle.get_raw());
+        let ty = TypeId::of::<M>();
+
+        let bgl = self.bgl.entry(ty).or_insert_with(|| {
+            mode.into_data(
+                || {
+                    let texture_binding = |idx: u32| BindGroupLayoutEntry {
+                        binding: idx as u32,
+                        visibility: ShaderStages::FRAGMENT,
+                        ty: BindingType::Texture {
+                            sample_type: TextureSampleType::Float { filterable: true },
+                            view_dimension: TextureViewDimension::D2,
+                            multisampled: false,
+                        },
+                        count: None,
+                    };
+
+                    let mut entries: Vec<_> = (0..texture_count).map(texture_binding).collect();
+                    entries.push(BindGroupLayoutEntry {
+                        binding: texture_count + 1,
+                        visibility: ShaderStages::VERTEX_FRAGMENT,
+                        ty: BindingType::Buffer {
+                            ty: BufferBindingType::Uniform,
+                            has_dynamic_offset: false,
+                            min_binding_size: NonZeroU64::new(data_count as _),
+                        },
+                        count: None,
+                    });
+                    device.create_bind_group_layout(&BindGroupLayoutDescriptor {
+                        label: Some("cpu material bgl"),
+                        entries: &entries,
+                    })
+                },
+                || {
+                    device.create_bind_group_layout(&BindGroupLayoutDescriptor {
+                        label: Some("gpu material bgl"),
+                        entries: &[BindGroupLayoutEntry {
+                            binding: 0,
+                            visibility: ShaderStages::VERTEX_FRAGMENT,
+                            ty: BindingType::Buffer {
+                                ty: BufferBindingType::Storage { read_only: true },
+                                has_dynamic_offset: false,
+                                min_binding_size: NonZeroU64::new(
+                                    (material.texture_count() + material.data_count()) as _,
+                                ),
+                            },
+                            count: None,
+                        }],
+                    })
+                },
+            )
+        });
+
+        self.type_info.insert(
+            ty,
+            PerTypeInfo {
+                data_count,
+                texture_count,
+            },
+        );
+
+        let translation_fn = texture_manager_2d.translation_fn();
 
         self.registry.insert(
             handle,
             InternalMaterial {
                 bind_group: mode.into_data(
                     || {
-                        BindGroupBuilder::new(None)
-                            .with_texture_view(material.albedo.to_texture(lookup_fn).unwrap_or(null_tex))
-                            .with_texture_view(material.normal.to_texture(lookup_fn).unwrap_or(null_tex))
-                            .with_texture_view(
-                                material
-                                    .aomr_textures
-                                    .to_roughness_texture(lookup_fn)
+                        let mut textures = vec![NonZeroU32::new(u32::MAX); texture_count as usize];
+                        material.to_texture(&mut textures, &mut translation_fn);
+
+                        let mut builder = BindGroupBuilder::new(None);
+                        for texture in textures {
+                            builder.append(BindingResource::TextureView(
+                                texture
+                                    .map(|tex| texture_manager_2d.get_view_from_index(tex))
                                     .unwrap_or(null_tex),
-                            )
-                            .with_texture_view(
-                                material
-                                    .aomr_textures
-                                    .to_metallic_texture(lookup_fn)
-                                    .unwrap_or(null_tex),
-                            )
-                            .with_texture_view(material.reflectance.to_texture(lookup_fn).unwrap_or(null_tex))
-                            .with_texture_view(
-                                material
-                                    .clearcoat_textures
-                                    .to_clearcoat_texture(lookup_fn)
-                                    .unwrap_or(null_tex),
-                            )
-                            .with_texture_view(
-                                material
-                                    .clearcoat_textures
-                                    .to_clearcoat_roughness_texture(lookup_fn)
-                                    .unwrap_or(null_tex),
-                            )
-                            .with_texture_view(material.emissive.to_texture(lookup_fn).unwrap_or(null_tex))
-                            .with_texture_view(material.anisotropy.to_texture(lookup_fn).unwrap_or(null_tex))
-                            .with_texture_view(material.aomr_textures.to_ao_texture(lookup_fn).unwrap_or(null_tex))
+                            ));
+                        }
+                        builder
                             .with_buffer(material_buffer.as_cpu())
-                            .build(device, self.bgl.as_cpu())
+                            .build(device, bgl.as_ref().as_cpu())
                     },
                     || (),
                 ),
@@ -368,44 +177,60 @@ impl MaterialManager {
         );
     }
 
-    pub fn update_from_changes(&mut self, queue: &Queue, handle: RawMaterialHandle, change: MaterialChange) {
-        let material = self.registry.get_mut(handle);
-        material.mat.update_from_changes(change);
+    // pub fn update_from_changes<M: MaterialTrait>(&mut self, queue: &Queue, handle: RawMaterialHandle, change: M) {
+    //     let material = self.registry.get_mut(handle);
+    //     material.mat.update_from_changes(change);
 
-        if let ModeData::CPU(ref mut mat_buffer) = material.material_buffer {
-            let cpu = CPUShaderMaterial::from_material(&material.mat);
-            queue.write_buffer(mat_buffer, 0, bytemuck::bytes_of(&cpu));
-        }
-    }
+    //     if let ModeData::CPU(ref mut mat_buffer) = material.material_buffer {
+    //         let cpu = CPUShaderMaterial::from_material(&material.mat);
+    //         queue.write_buffer(mat_buffer, 0, bytemuck::bytes_of(&cpu));
+    //     }
+    // }
 
-    pub fn get_material(&self, handle: RawMaterialHandle) -> &Material {
-        &self.registry.get(handle).mat
-    }
+    // pub fn get_material(&self, handle: RawMaterialHandle) -> &Material {
+    //     &self.registry.get(handle).mat
+    // }
 
-    pub fn get_bind_group_layout(&self) -> &BindGroupLayout {
-        self.bgl.as_ref().into_common()
-    }
+    // pub fn get_bind_group_layout(&self) -> &BindGroupLayout {
+    //     self.bgl.as_ref().into_common()
+    // }
 
-    pub fn cpu_get_bind_group(&self, handle: RawMaterialHandle) -> (&BindGroup, SampleType) {
-        let material = self.registry.get(handle);
-        (material.bind_group.as_cpu(), material.mat.sample_type)
-    }
+    // pub fn cpu_get_bind_group(&self, handle: RawMaterialHandle) -> (&BindGroup, SampleType) {
+    //     let material = self.registry.get(handle);
+    //     (material.bind_group.as_cpu(), material.mat.sample_type)
+    // }
 
-    pub fn gpu_get_bind_group(&self) -> &BindGroup {
-        self.bg.as_gpu()
-    }
+    // pub fn gpu_get_bind_group(&self) -> &BindGroup {
+    //     self.bg.as_gpu()
+    // }
 
-    pub fn internal_index(&self, handle: RawMaterialHandle) -> usize {
-        self.registry.get_index_of(handle)
-    }
+    // pub fn internal_index(&self, handle: RawMaterialHandle) -> usize {
+    //     self.registry.get_index_of(handle)
+    // }
 
     pub fn ready(&mut self, device: &Device, queue: &Queue, texture_manager: &TextureManager) {
         profiling::scope!("Material Ready");
-        self.registry.remove_all_dead(|_, _, _| ());
+        self.registry.remove_all_dead();
 
         if let ModeData::GPU(ref mut buffer) = self.buffer {
             profiling::scope!("Update GPU Material Buffer");
             let translate_texture = texture_manager.translation_fn();
+
+            let count: usize = self
+                .registry
+                .architype_lengths()
+                .map(|(ty, len)| {
+                    let type_info = self.type_info[&ty];
+
+                    len * (round_up_pot(type_info.texture_count, 16) + round_up_pot(type_info.data_count, 16))
+                })
+                .sum();
+            
+            let mut data = vec![0u8; count];
+            
+
+            
+
             let data: Vec<_> = self
                 .registry
                 .values()
@@ -425,4 +250,20 @@ fn create_gpu_buffer_bg(device: &Device, bgl: &BindGroupLayout, buffer: &Buffer)
     BindGroupBuilder::new(Some("gpu material bg"))
         .with_buffer(buffer)
         .build(device, bgl)
+}
+
+fn write_gpu_materials<M: MaterialTrait>(mut dest: &mut [u8], vec_any: &VecAny, translation_fn: &mut (dyn FnMut(&TextureHandle) -> NonZeroU32 + '_)) {
+    let materials = vec_any.downcast_slice::<ArchitypeResourceStorage<M>>().unwrap();
+
+    for mat in materials {
+        let mat_size = round_up_pot((mat.data.texture_count() * 4) as usize, 16);
+        mat.data.to_texture(bytemuck::cast_slice_mut(&mut dest[0..mat_size]), translation_fn);
+
+        dest = &mut dest[mat_size..];
+
+        let data_size = round_up_pot(mat.data.data_count() as usize, 16);
+        mat.data.to_data(&mut dest[0..data_size]);
+
+        dest = &mut dest[data_size..];
+    }
 }
